+++ project/Dependencies.scala
@@ -46,7 +46,8 @@ object Dependencies {
   val googleIam: ModuleID =                  "com.google.apis"       % "google-api-services-iam"                  % s"v1-rev215-$googleV"
   val googleBigQuery: ModuleID =             "com.google.apis"       % "google-api-services-bigquery"             % s"v2-rev377-$googleV"
   val googleGuava: ModuleID = "com.google.guava"  % "guava" % "22.0"
-  val googleRpc: ModuleID =               "io.grpc" % "grpc-core" % "1.5.0"
+  val googleRpc: ModuleID =               "io.grpc" % "grpc-core" % "1.15.0"
+  val googleFirestore: ModuleID = "com.google.cloud" % "google-cloud-firestore" % "0.70.0-beta"
 
   val commonDependencies = Seq(
     scalaLogging,
@@ -91,6 +92,8 @@ object Dependencies {
     googleBigQuery,
     googleGuava,
     googleRpc,
+    googleFirestore,
+    catsEffect,
     akkaHttpSprayJson,
     akkaTestkit
   ).map(excludeGuavaJDK5)