+++ google/CHANGELOG.md
@@ -4,10 +4,12 @@ This file documents changes to the `workbench-google` library, including notes o
 
 ## 0.18
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.18-7e86fba"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.18-TRAVIS-REPLACE-ME"`
 
 ### Changed
 - Moved `SamModel`
+- add googleFirestore support
+- add DistributedLock backed by google firestore
 
 ## 0.17
 - upgrade cats to 1.4.0 and scala to 2.12.7